requirements.txt
openai-whisper==20231117
openai==1.12.0
moviepy==1.0.3
python-dotenv==1.0.0 